I WANNA IED

+++ ft_strlen.c
@@ -1,8 +1,6 @@
-#include "libft.h"
-
-size_t ft_strlen(const char *s)
+int ft_strlen(const char *s)
 {
-	size_t i;
+	int i;
 
 	i = 0;
 	while (s[i])
